move resource => included resource conversion to lib/jsonapi
